% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parkinsons.R
\docType{data}
\name{parkinsons}
\alias{parkinsons}
\title{Network meta-analysis in Parkison's disease}
\format{
A data frame with 15 rows and 6 variables:
\tabular{rl}{
\bold{\emph{studyn}}\tab Study name \cr
\bold{\emph{trtn}}\tab Treatment name \cr
\bold{\emph{y}}\tab Mean-off time reduction (hours) \cr
\bold{\emph{se}}\tab Standard error \cr
\bold{\emph{diff}}\tab Mean differences vs reference arm in off time reduction (hours) \cr
\bold{\emph{se_diff}}\tab Standard error of mean difference \cr
}
}
\usage{
data(parkinsons)
}
\description{
Data frame containing the mean off-time reduction
in patients given dopamine agonists as adjunct therapy
in Parkinson's disease, from 7 trials comparing four
active drugs and placebo (Dias et al. 2011).
}
\examples{
data(parkinsons)
head(parkinsons)
}
\references{
Dias S, Welton NJ, Sutton AJ, Ades AE (2011). “NICE DSU Technical Support
Document 2: A generalised linear modelling framework for pair-wise and network meta-analysis
of randomised controlled trials.” National Institute for Health and Care Excellence.
https://www.sheffield.ac.uk/nice-dsu.
}
\keyword{datasets}
